make schedule its own tab, reuse across both pages
