+++ apps/frontend/pages/admin/event/[eventId]/division/[divisionId].tsx
@@ -1,20 +1,14 @@
 import { GetServerSideProps, NextPage } from 'next';
-import { TabContext, TabPanel } from '@mui/lab';
-import { Paper, Tabs, Tab, Stack } from '@mui/material';
+import { Paper, Stack } from '@mui/material';
 import { WithId } from 'mongodb';
-import dayjs from 'dayjs';
 import { FllEvent, AwardSchema, Division } from '@lems/types';
 import { apiFetch, serverSideGetRequests } from '../../../../../lib/utils/fetch';
 import Layout from '../../../../../components/layout';
-import GenerateScheduleButton from '../../../../../components/admin/generate-schedule';
 import EditDivisionForm from '../../../../../components/admin/edit-division-form';
-import DivisionAwardEditor from '../../../../../components/admin/division-award-editor';
-import DeleteDivisionData from '../../../../../components/admin/delete-division-data';
-import DivisionScheduleEditor from '../../../../../components/admin/division-schedule-editor';
 import DownloadUsersButton from '../../../../../components/admin/download-users';
 import UploadFileButton from '../../../../../components/general/upload-file';
+import DivisionTabEditor from '../../../../../components/admin/division-tab-editor';
 import { localizeDivisionTitle } from '../../../../../localization/event';
-import { useQueryParam } from '../../../../../hooks/use-query-param';
 
 interface Props {
   event: WithId<FllEvent>;
@@ -23,59 +17,24 @@ interface Props {
 }
 
 const Page: NextPage<Props> = ({ event, division, awardSchema }) => {
-  const [activeTab, setActiveTab] = useQueryParam('tab', '1');
-
   return (
     <Layout
-      maxWidth="md"
+      maxWidth="lg"
       title={`ניהול אירוע: ${localizeDivisionTitle({ ...division, event })}`}
       back={`/admin/event/${event._id}`}
     >
-      <TabContext value={activeTab}>
-        <Paper sx={{ mt: 2 }}>
-          <Tabs
-            value={activeTab}
-            onChange={(_e, newValue: string) => setActiveTab(newValue)}
-            centered
-          >
-            <Tab label="פרטי הבית" value="1" />
-            <Tab label="מסגרת אירוע" value="2" />
-            <Tab label="פרסים" value="3" />
-          </Tabs>
-        </Paper>
-        <TabPanel value="1">
-          <Stack spacing={2}>
-            <EditDivisionForm event={event} division={division} />
-            <Paper sx={{ p: 4 }}>
-              {division?.hasState && <DeleteDivisionData division={{ ...division, event }} />}
-              <Stack justifyContent="center" direction="row" spacing={2}>
-                <UploadFileButton
-                  urlPath={`/api/admin/divisions/${division?._id}/schedule/parse`}
-                  displayName="לוח זמנים"
-                  extension=".csv"
-                  disabled={division?.hasState}
-                  requestData={{ timezone: dayjs.tz.guess() }}
-                />
-                <GenerateScheduleButton division={division} />
-                <DownloadUsersButton division={division} disabled={!division?.hasState} />
-              </Stack>
-            </Paper>
-            <Paper sx={{ p: 4, display: 'flex', justifyContent: 'center' }}>
-              <UploadFileButton
-                urlPath={`/api/admin/divisions/${division?._id}/pit-map`}
-                displayName="מפת פיטים"
-                extension=".png"
-              />
-            </Paper>
-          </Stack>
-        </TabPanel>
-        <TabPanel value="2">
-          <DivisionScheduleEditor event={event} division={division} />
-        </TabPanel>
-        <TabPanel value="3">
-          <DivisionAwardEditor divisionId={division?._id} awardSchema={awardSchema} />
-        </TabPanel>
-      </TabContext>
+      <EditDivisionForm event={event} division={division} sx={{ mt: 2, p: 4 }} />
+      <Paper sx={{ p: 4, mt: 2 }}>
+        <Stack direction="row" spacing={2} alignItems="center" justifyContent="center">
+          <UploadFileButton
+            urlPath={`/api/admin/divisions/${division?._id}/pit-map`}
+            displayName="מפת פיטים"
+            extension=".png"
+          />
+          <DownloadUsersButton division={division} disabled={!division?.hasState} />
+        </Stack>
+      </Paper>
+      <DivisionTabEditor event={event} division={division} awardSchema={awardSchema} />
     </Layout>
   );
 };
